convert int to long

+++ features/merchant/seller_menu_common/src/main/java/com/tokopedia/seller/menu/common/domain/usecase/GetShopTotalFollowersUseCase.kt
@@ -10,7 +10,7 @@ import com.tokopedia.seller.menu.common.domain.entity.ShopTotalFollowers
 import com.tokopedia.usecase.coroutines.UseCase
 import javax.inject.Inject
 
-class GetShopTotalFollowersUseCase @Inject constructor(private val gqlRepository: GraphqlRepository) : UseCase<Int>() {
+class GetShopTotalFollowersUseCase @Inject constructor(private val gqlRepository: GraphqlRepository) : UseCase<Long>() {
 
     companion object {
         const val QUERY = "query SingleShopTotalFollowers(\$shopID: Int!) {\n" +
@@ -37,7 +37,7 @@ class GetShopTotalFollowersUseCase @Inject constructor(private val gqlRepository
 
     var params = HashMap<String, Any>()
 
-    override suspend fun executeOnBackground(): Int {
+    override suspend fun executeOnBackground(): Long {
         val gqlRequest = GraphqlRequest(QUERY, ShopTotalFollowers::class.java, params)
         val cacheStrategy = GraphqlCacheStrategy.Builder(CacheType.CLOUD_THEN_CACHE)
                 .build()
